optimizations, anti-silent hysteresis, fast-track
anti-silent hysteresis -- we'll trigger an immediate transcription
if the user goes silent, but now only once in a SILENT_TIMEOUT period.

Before we would trigger a transcription every time the user went
silent, which could have been several times a second for users
at the edge of the transmit threshold.

fast-track -- look ahead at the audio that's been processed
while we're waiting for the transcription to complete.  If
enough time has passed to know what happened, and if we don't
see anything but silence in the interval immediately after the
transcription, just go ahead and finalize the entire
transcription immediately.