create "transcript_strategy" trait, replacing "transcript_director"
the transcript_strategy determines:
 - when to request a new transcript
 - how much of that transcript to finalize

This replaces the previous "transcript_director".

This isn't quite done but works and is quite the improvement,
so checking in for safekeeping.

The advantage here is that all of the "magic" for determining
when to request a new transcript is now in one place, which
greatly simplifies everything else around it.