Rewrite (#1)
* wip rewrite

* safekeeping commit

* safekeeping checkin

* tests for call_attendees

* safekeeping

* finish voice_activity

* fix warnings

* clippy auto-fixes

* more clippy fixes

* safekeeping

* fix clippy issues

* safekeeping merge

* safekeeping

* safekeeping

* safekeeping

* safekeeping

* clippy auto-fixes

* last cliippy fixes

* safekeeping

* clippy fixes

* safekeeping... progress, slow but there

* clippy fixes

* safekeeping (not building)

* make things build, getting closer!

* fix clippy

* make queue names more consistent

* more fixes

* safekeeping, almost all connected!

* remove unused constants

* fix loop, clippy

* safekeeping

* clippy

* wip, not building

* safekeeping... progress!

* clippy fixes

* notes to future self

* safekeeping... maybe complete?

* clippy

* more clippy

* add test

* add tests

* more tests

* move lots of files around

* clippy

* keep list of recent tokens

* clippy fixes

* move moving

* renaming

* fix for cleanup bug

* clippy fixes

* logging for lots of problems

* logging, cleanup

* some progress

* send notice when user stops talking

to trigger an eager transcription.  This is in addition
to the existing trigger when the user stops talking
for 2 seconds.

* more progress, still double-sampling audio

* kinda works, almost!

* only use single slice

* lots more fixes

* minor fixes

* clippy auto-fixes

* manual clippy fixes

* fix tests

* ignore silent audio when voice not active

* get rid of Bytes to try and get rid of a strange bug

* optimizations, anti-silent hysteresis, fast-track

anti-silent hysteresis -- we'll trigger an immediate transcription
if the user goes silent, but now only once in a SILENT_TIMEOUT period.

Before we would trigger a transcription every time the user went
silent, which could have been several times a second for users
at the edge of the transmit threshold.

fast-track -- look ahead at the audio that's been processed
while we're waiting for the transcription to complete.  If
enough time has passed to know what happened, and if we don't
see anything but silence in the interval immediately after the
transcription, just go ahead and finalize the entire
transcription immediately.

* fold user_audio into user_slice

* split transcript-generating logic into transcript_director

refactors logic related to transcript generation to
a new struct, TranscriptDirector, which is responsible
for determining when to request transcripts, and when
to report the results to the API.

The audio_slice class is now only responsible for storing
audio in the correct location of the buffer.

* move is_interval_silent into audio_slice

* improve tests

* print transcription lag

* return to using Bytes

* clippy auto-fixes

* exit properly

* new directory structure

* rename AudioSlice to AudioBuffer

* move to general event class

to support future refactoring

* split out async audio worker

* split audio worker into its own file
rename audio mux back to audio manager
fix test flake, hopefully

* create "transcript_strategy" trait, replacing "transcript_director"

the transcript_strategy determines:
 - when to request a new transcript
 - how much of that transcript to finalize

This replaces the previous "transcript_director".

This isn't quite done but works and is quite the improvement,
so checking in for safekeeping.

The advantage here is that all of the "magic" for determining
when to request a new transcript is now in one place, which
greatly simplifies everything else around it.

* create "strategies" dir

also, separate implementation of channel mute detector

* use priority queue for idle detection

a few more lines of code, but cleaner

* implement 5-second strategy

 - kick off translation after 5 seconds, or
   silence interval
 - keep any segments that end more than 1 second
   at the end of the analyzed audio
 - keep all segments if the following 1 second of audio
   is totally silent

* clippy auto-fixes

* manual clippy fixes