improved the five second strategy to a 5/1 second strategy
We'll now request our first transcription at 5 seconds, and
then every subsequent transcription at 1 second intervals.

Clearing audio the audio buffer resets the strategy, and
partially removing the audio is accounted for.  The calculation
is based on how much audio remains in the buffer.

Also, a fix where we would request a new transcription while
we were still waiting for the previous transcription to complete.
Just add a blanket check to make sure this can't happen.