optimization: filter out low-energy signals before whisper
When in a chat with multiple people with poor microphones,
they'll often always send short bursts of low-energy data.

This is a problem because transcoding these is expensive,
and with n people always in this state we need to be able handle
transcriptions at a speed of 1/n realtime.  But this isn't normally
necessary, since usually only 1 or 2 people will be talking, so
provisioning for that is wasteful.

This change sets a minimum threshold for the energy of a
signal before we will send it to whisper.  This is sets
by DONT_EVEN_BOTHER_RMS_THRESHOLD, but should be made
configurable in the future.

Also, print debug messages for the energy of the signal
that we are taking, so we can see if the threshold is
set reasonably.